implemented more UserHome features

+++ TCGTradingPost/src/main/java/com/tcgtp/repositories/OrderStatusRepository.java
@@ -4,6 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.tcgtp.domain.OrderStatus;
 
-public interface OrderStatusRepository extends CrudRepository<OrderStatus, Integer>{
-
+public interface OrderStatusRepository extends CrudRepository<OrderStatus, Long>{
 }
